Splitted EDB Unit Tests to speedup CICD
